requirements.txt
schedule
python-dotenv
secure-smtplib
